+++ app/controllers/users/sessions_controller.rb
@@ -1,37 +1,18 @@
-# frozen_string_literal: true
-
 class Users::SessionsController < Devise::SessionsController
-  # before_action :configure_sign_in_params, only: [:create]
 
-  # GET /resource/sign_in
-  # def new
-  #   super
-  # end
-
-  # POST /resource/sign_in
-   def create
-
+  def create
     super
+    return if !user_signed_in?
     @cart = current_user.cart
-
     map_session_items_to_cart
-    delteCartSeesion
+    delete_cart_session
     take_items_into_session
-  
-   end
+  end
 
-  # DELETE /resource/sign_out
-   def destroy
-    delteCartSeesion
+  def destroy
+    delete_cart_session
     super
-   end
-
-  # protected
-
-  # If you have extra params to permit, append them to the sanitizer.
-  # def configure_sign_in_params
-  #   devise_parameter_sanitizer.permit(:sign_in, keys: [:attribute])
-  # end
+  end
 
   private
 
@@ -40,10 +21,7 @@ class Users::SessionsController < Devise::SessionsController
     
     return if items.nil?
 
-    if !session[:cart]
-      session[:cart]=Array.new
-      session[:cart_id]=@cart.id
-    end
+    session[:cart]=Array.new if !session[:cart]
     
     items.each do |i|
       session[:cart].push(i)
@@ -52,24 +30,16 @@ class Users::SessionsController < Devise::SessionsController
 
   def map_session_items_to_cart
     if session[:cart]
-      hash=session[:cart]
-      if @cart.nil?
-        @cart=current_user.build_cart
-      end
-      hash.each do |o|
+      @cart=current_user.build_cart if @cart.nil?
+      
+      session[:cart].each do |o|
         Item.find(o['id']).update(purchaseable: @cart)
       end
     end
   end
   
-  def delteCartSeesion # correct these spellings
-    if session[:cart]
-      session.delete(:cart)
-     end
-
-     if session[:cart_obj]
-      session.delete(:cart_obj) # review here
-     end
+  def delete_cart_session
+    session.delete(:cart) if session[:cart]
   end
 
 end
